remove wrapper.*
The OSGi Container (Concierge) is controlled by the wrapper.
Since OSGi bundles (the node is such a thing now) don't have
access to the 'outer world' per default the updating stuff
needs a redesign.
The shut down signaling (isShutDownInProgress()) is done by
the bundle.
So the old stuff is left as comments to fix it later